requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=1.0.0
git+https://github.com/openai/CLIP.git
plyfile>=1.0.0
POT>=0.9.0
lapjv>=1.3.0
scipy>=1.11.0
numpy>=1.24.0
pyyaml>=6.0
wandb>=0.16.0
tqdm>=4.66.0
